requirements.txt
line_bot_sdk==1.19.0
beautifulsoup4==4.9.3
Flask==1.1.2
gunicorn==20.1.0
